+++ src/fractal/lib/global_bindings.wgsl
@@ -24,9 +24,9 @@ struct VertexOutput {
 @group(0) @binding(0)
 var<uniform> globals: GlobalUniforms;
 
-fn max_iterations() -> f64
+fn max_iterations() -> f32
 {
-    return f64(globals.max_iterations)*max(1.0, log(f64(globals.zoom)));
+    return f32(globals.max_iterations)*max(1.0, log(f32(globals.zoom)));
 }
 fn view_radius() -> f32
 {
@@ -37,17 +37,17 @@ fn epsilon() -> f32
     return 0.00000000001;
 }
 
-fn z_in(position: vec4<f32>) -> vec2<f64>
+fn z_in(position: vec4<f32>) -> vec2<f32>
 {
-    let pos = vec2(f64(position.x), f64(position.y))/f64(position.w) - vec2(f64(globals.window_size.x), f64(globals.window_size.y))/2.0;
-    return cmul(pos/f64(globals.zoom), cis(f64(globals.rot))) - vec2(f64(globals.center.x), f64(globals.center.y));
+    let pos = vec2(f32(position.x), f32(position.y))/f32(position.w) - vec2(f32(globals.window_size.x), f32(globals.window_size.y))/2.0;
+    return cmul(pos/f32(globals.zoom), cis(f32(globals.rot))) - vec2(f32(globals.center.x), f32(globals.center.y));
 }
 
-fn shift_in() -> vec2<f64>
+fn shift_in() -> vec2<f32>
 {
-    return vec2(f64(globals.shift.x), f64(globals.shift.y));
+    return vec2(f32(globals.shift.x), f32(globals.shift.y));
 }
-fn exp_in() -> vec2<f64>
+fn exp_in() -> vec2<f32>
 {
-    return vec2(f64(globals.exp.x), f64(globals.exp.y));
+    return vec2(f32(globals.exp.x), f32(globals.exp.y));
 }
