+++ src/fractal/supergolden_julia.wgsl
@@ -29,7 +29,7 @@ fn fs_main(@builtin(position) position: vec4<f32>) -> @location(0) vec4<f32>
     {
         z = cmul(z + c, powc(z, e)) + c;
     }
-    let m = f32(f64(i) - log(log(norm(z)))/log(norm(e)));
+    let m = f32(f32(i) - log(log(norm(z)))/log(norm(e)));
     let zz = vec2(f32(z.x), f32(z.y));
 
     return colormap3(zz, m);
